Refactor Builder class into PolygonBuilder
